Add files via upload
ActualitéWeb is a powerful and versatile news aggregator web application that enables users to stay up-to-date with the latest news from various sources. Built with Node.js, TypeScript, and Postgres, this application fetches news articles using web scraping and RSS feeds and presents them in a simple and responsive interface that is easy to use.

The application boasts a wide range of features that make it a valuable tool for anyone looking to stay informed. These include search functionality, user authentication, news categories, recommendations, and social media sharing. With its intuitive design and user-friendly interface, ActualitéWeb is an ideal solution for anyone who wants to keep track of the latest news without having to navigate multiple websites and news sources.

The developers of ActualitéWeb are committed to ongoing improvement and innovation, with plans to further enhance the user interface and integrate with external APIs. This commitment to excellence is reflected in the application's open-source nature, which allows anyone to contribute to the development of the project.

The application is licensed under the Apache License 2.0, which provides users with the freedom to use, modify, distribute, and sell the software without any warranty or liability from the original author or contributor. However, any modifications or distributions must be licensed